requirements.txt
django-environ==0.9.0
psycopg2-binary==2.9.5
django-silk==5.0.2
django-mptt==0.14.0
django-nested-inline==0.4.5
aiogram==2.23.1
black==22.10.0
redis==4.4.0
aioredis==2.0.1
celery==5.2.7
django-celery-beat==2.4.0
pillow==9.3.0
django_ajax_datatable==4.4.4
charset-normalizer==2.0.10
cryptography==36.0.1
defusedxml==0.7.1
Django==4.1.4
django-crispy-forms==1.13.0
django-otp==1.1.3
idna==3.3
pycparser==2.21
python3-openid==3.2.0
pytz==2021.3
qrcode==7.3.1
requests==2.27.1
opencv-python==4.6.0.66
